+++ src/encoder_avx_base.h
@@ -39,7 +39,7 @@ template<enum YEncDecIsaLevel use_isa>
 static size_t do_encode_avx2(int line_size, int* colOffset, const unsigned char* src, unsigned char* dest, size_t len) {
 	unsigned char* es = (unsigned char*)src + len;
 	unsigned char *p = dest; // destination pointer
-	long i = -len; // input position
+	long i = -(long)len; // input position
 	unsigned char c, escaped; // input character; escaped input character
 	int col = *colOffset;
